# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Ban:
    properties:
      autoBan:
        type: boolean
      banID:
        type: integer
      banReason:
        type: string
      bannedAt:
        type: string
      bannerID:
        type: string
      guildID:
        type: string
      memberID:
        type: string
    type: object
  models.Channel:
    properties:
      channelID:
        type: string
      guildID:
        type: string
      ignored:
        type: boolean
      wpBlacklisted:
        type: boolean
    type: object
...
host: cardinal.gyroskan.com
info:
  contact:
    email: gyroskan@gmail.com
    name: API Support
  description: The API to interact with cardinal discord bot database.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Cardinal API
  version: "0.1"
paths:
...
  /guilds/{guildID}/channels:
    get:
      description: Fetch all channels of the guild.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - default: false
        description: ignored channels only
        in: query
        name: ignored
        type: boolean
      - default: false
        description: xpBlacklist channels only
        in: query
        name: xpBlacklist
        type: boolean
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "500":
          description: Server error
      summary: Get Guild channels
      tags:
      - Channels
    post:
      consumes:
      - application/json
      description: Create a new channel for a guild.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: Channel values
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/models.Channel'
      produces:
      - application/json
      responses:
        "201":
          description: Created channel
          schema:
            $ref: '#/definitions/models.Channel'
        "400":
          description: Wrong values
        "403":
          description: Forbidden
        "500":
          description: Server Error
      summary: Create channel
      tags:
      - Channels
  /guilds/{guildID}/channels/{channelID}:
    delete:
      consumes:
      - application/json
      description: Delete a guild channel
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: Channel id
        in: path
        name: channelID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "206":
          description: No Content
        "403":
          description: Forbidden
        "404":
          description: Not Fountd
        "500":
          description: Server Error
      summary: Delete guild channel
      tags:
      - Channels
    get:
      description: Fetch the channel of the guild.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: channel id
        in: path
        name: channelID
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Server error
      summary: Get one Guild Channel
      tags:
      - Channels
    patch:
      consumes:
      - application/json
      description: Update fields of a guild's channel
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: Channel id
        in: path
        name: channelID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: Not Fountd
        "500":
          description: Server Error
      summary: Update channel values
      tags:
      - Channels
...
  /guilds/{guildID}/members/{memberID}/bans:
    get:
      description: Fetch all bans of the member.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: member id
        in: path
        name: memberID
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "500":
          description: Server error
      summary: Get Member Bans
      tags:
      - Bans
    post:
      consumes:
      - application/json
      description: Create a new ban for a member.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: member id
        in: path
        name: memberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created role
          schema:
            $ref: '#/definitions/models.Ban'
        "400":
          description: Wrong values
        "403":
          description: Forbidden
        "500":
          description: Server Error
      summary: Create ban
      tags:
      - Bans
  /guilds/{guildID}/members/{memberID}/bans/{banID}:
    delete:
      consumes:
      - application/json
      description: Delete a member's ban
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: member id
        in: path
        name: memberID
        required: true
        type: string
      - description: ban id
        in: path
        name: banID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "206":
          description: No Content
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Server Error
      summary: Delete member's ban
      tags:
      - Bans
    get:
      description: Fetch the ban of the member.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: member id
        in: path
        name: memberID
        required: true
        type: string
      - description: ban id
        in: path
        name: banID
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Server error
      summary: Get one ban
      tags:
      - Bans
...
  /users/{username}/ban:
    delete:
      description: Update User ban. POST to unbann, DELETE to ban.
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: NotFound
        "500":
          description: Server error
      summary: Ban User
      tags:
      - Users
    post:
      description: Update User ban. POST to unbann, DELETE to ban.
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: NotFound
        "500":
          description: Server error
      summary: Ban User
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"